Add duplicate detection for new task
There may be instances where users mistakenly input the same tasks multiple times.
With duplicates in the task list, it may cause confusion for the user, especially when
there are no dates associated to it (e.g. todo tasks).

Let's add a duplicate detection method during the addition of a new task.

This will prevent tasks of duplicate descriptions being in the list at the same time.