Add assertions to improve code robustness
This commit adds assertions to the Roo class so that potential issues
can be catch earlier during program execution.

This is because some input is assumed to be not null and assertions
help to check for potential mistakes that may accidentally set them
to null.

Let's add assertions to check that the "filepath" argument in the
constructor and the "input" argument in the "run" method is not null.

These changes will help in identifying and debugging issues more easily
during development.

Refer to this website on dealing with assertions
https://www.geeksforgeeks.org/assertions-in-java/