Improved Code Quality of James
This commit improved the readability of the messages from James after each command. Previously, the strings were added but now
they are written in a single line which helps make it more readable.

This commit also includes comments for the different types of commands to make the parsing easier to understand.