Add sort tasks alphabetically feature
Currently, there is no feature to sort tasks in alphabetical order.

Sorting tasks according to alphabetical order can enhance visibility for the user.

It is done through a method sortListAlphabetically so that it is in the same command format as the other commands.

It is done this way so that users can easily learn how to use the command.