Add tag to the chatbot, ability to view tag, delete tag
Add feature to include tags now with task, tags are now stored alongside
with the task.

Used common logic for marking where the list is regenerated each time
after change of tags, can be possibly optimize later on.

In this case, design decision was to have a separate list for showing
tags that did not show their timings if they do have. This is to
prevent cluttering and better user experience for the user

Tested the launcher for functionality. Fixed data corruption problem
earlier where storage containing information would crash. Problem was
noted to be out of bounds error due to wrong checking logic