Commandable: Refactor naming to Executable
Commandable is not a very suitable name for this interface, since it is not a proper english word and does not represent what it does. Since the shared method is named execute(), this schema makes more sense.

Replace all instances of the word "Commandable" with the word "Executable", for a more accurate representation of the name.