Improve code quality
There is a very long method which is parse method in Parse class.
Refer to the SLAP, this method having multiple levels of
abstraction.

I extract the long method to several small methods.

By doing this this method succesfully follow the SLAP,
and the readability is increased.

Beside, I also change some of the variable names to be consistent,
when these variables inside the long method, there cannot be
more than one variable with the same name.