Add assertions to key places in the code
Currently, there is no assertions put in place to document important
assumptions in the code.

Assertions is important to ensure workability of code and to
reveal any potentials bugs in the logic.

Add assertions to to certain places in the code, mainly in Aj.java

Most of the key logic and assumption is in Aj.java, as such most
assertions are done there.