Use assert to enforce FileStorage assumptions
FileStorage currently does not enforce the dataFilePath and has a
slightly awkward way of handling missing keys in the parsed data.

This runs the risk of having invalid input and the use of
IllegalStateException does not make a lot of sense in the long run.

Adopt assert statements to ensure dataFilePath is not empty and
AssertionError is thrown instead of IllegalStateException for missing
keys.

This improves the visibility of configuration issues and makes it
clearer why exactly the bot failed.