Add assertion commands
There are several sections of the code where assumptions about the program state are made.

Adding these assertions can ensure that when the execution comes to the assert command, these assumptions are verified by the runtime.

Let's add assertions in
* Duke class to ensure the TaskList is intialised before running application.
* Mainwindow class to make sure that a String is return from a getresponse() before creating Diaglog object.
* Parser class to ensure that the userInput parsed in is not empty and if-statement elimates all cases where the keyword is empty.
* Tasklist class to ensure the tasks is successfully added, marked and unmarked or that else statement handles Event tasks.

assertions ensure the assumptions made are correct and that no mistake in the code goes unnoticed.