Change Parser to adhere to code quality guideline
The Parser.java file contains a parse() method that parses the commands
input by the user into the chatbot. However, it is a very long method
and could be split up.

The if-else syntax is replaced by a switch-case syntax with minor
modifications to the logic to accommodate this syntax (should not be
noticed by users with normal use). This improves code readability.

Additionally, most multi-statement code is replaced by more private
static methods within the Parser class to ease readability of the
parse() method even more.

This fix in code quality was inspired and spearheaded by the comments
in this PR: https://github.com/nus-cs2103-AY2324S1/ip/pull/288