New Unit Tests: ParserTest.java created for unit testing on Parser class
The getCommand method from the Parser class did not have unit tests
created for it. It had unexpected behaviors when the user inputs their
command with a trailing whitespace.

Let's
* Create unit tests for the getCommand method to ensure future changes
will not cause regressions for this command.
* Create unit tests for the getCommand method to ensure the code is able
to handle trailing whitespaces in the user input.

Allowing for proper handling of trailing whitespaces will allow more
flexibility in the user input, allowing the chatbot to handle more typo
mistakes made by the user.