Add assert statements to some classes
There are several assumptions that are made when executing certain
steps in the code.

Assertions are used to define assumptions about the program state so
that the runtime can verify them.

As a step towards verifying those assumptions, lets add some assert
statements. Doing so will make it easier to find possible bugs in
the code.