Add assertions to files
Directory path in Duke and execute methods in Command classes may
contain null variables.

Null variables may cause the program to become illogical and crash.

Assertions are added to check if these relevant variables are null
at runtime and if they are, the program will exit.

Assertions help to ensure that the assumptions stated in the code
about the program will remain adhered to throughout runtime.