Refactor LocalDateTime formatting
Formatting for LocalDateTime is expected to be consistent for the entire
program when displayed to the user in the GUI.

By introducing a new Formatter utility class with a single static method
formatDateTime, it reduces the amount of clutter within the methods
in which they are used. This also ensures consistency in the formatting
of LocalDateTime as String objects within the GUI.

In addition, removed unnecessary this.* references from methods when
referring to instance fields.