Update build.gradle file
The main class in the build.gradle file is referring to the duke.java

Implementation of the duke application now relies on JavaFX.
duke.Launcher is the file responsible for launching the GUI and the
Duke application as a whole. Running the current JAR file does not
cause the GUI the pop up.

Let's change the main class in the build.gradle file to duke.Launcher.

The JAR file created after this change will reflect the GUI to
appear, which represents the current version of the duke
application.

Some unused code was removed as well in this commit.