Command classes: Write assertions for `execute` method
The `execute` method in some command classes require the command objects
to be correctly initialized. This means that some fields must not be
empty, such as the "by" clause for the `Deadline` objects, and "from"
and "to" clauses for the `Event` objects. The cases where the user
forgets to fill in these clauses should already be handled. These
assertions ensure that future changes to the code base will still allow
them to be handled.

Let's:
* Create `assert` statements for the required fields to be not `null` to
allow the `execute` method to function as expected.

Assertions will ensure that the fields will not be left as empty in the
event that the exceptions fail to be raised.