Add assertions to Parser and MainWindow
To ensure that certain private methods will only be run in certain
scenarios, assertions are used to enforce a precondition at the start
of

Parser class: Ensure that the different commands parsed, as handled by
different private methods within the Parser class only happen when the
message starts with the corresponding command message. For example,
only inputs starting with 'mark' can trigger the parseMarkCommand()
private static method in Parser and return a MarkCommand object.

MainWindow class: Ensure that the handleExit() method is only run if the
user specifies and types "bye". Checks if the input message starts with
"bye".

Gradle is also configured to handle assertions in this commit.