Added java assertions in some classes
An assertion was added at the end of runInput method (Parser) since it
is not intended for the method to reach the end. One was also added for
retrieveTask method (TaskList) as it cannot be called if the length
of the arrayList is 0. For mark and unmark, we also want to ensure that
the correct completion status is printed. To do this, let's add an
assertion to check the boolean isDone.