Implement Sort feature for tasks
Previously, the user does not have any reasonable way of sorting the
tasks that they have added other than to edit the datafile directly,
which may be dangerous as erroneous edits may cause the program to
crash.

Adding a sort feature ensures that the user is now able to better
organise their tasks in a ascending/descending order, based on the
task description, start date, end date and deadline of tasks.

To prevent user-induced errors due to careless edits of the datafile,
as well as to give users more functionality to customise their
task lists, let's give users a convenient but powerful sort feature
to allow them to do so.