Add assertions to DeleteCommand.
DeleteCommand can only occur if the task list is not empty.

Adding assertion to ensure that task list is not empty ensures that the
DeleteCommand is only called when there is something to delete.

Let's add the assertion that the task list is not null to make sure that
the assumption is fulfilled before command is executed.