Provide a way to get reminders about task
Current Situation
Deadline Class do not provide a method to check if the deadline of a particular Task is within x days.

Why it needs to change
This allows users to make full use of the functionality of the DUKE application as well as the special LocalDate object in Deadline class.

What is being done about it
Deadline class has a method that checks if a particular Deadline Task's due date is within x days. Parser Class also has the appropriate conditional checks to ensure users prompts are correct

Why it is done that way
This ensures proper abstraction