Add assertions to Parser
There are a few places in Parser where after splitting a string, the length of the resultant array is checked (and it should be 2).

Adding assertions in these places can help make explicit the length of the resultant arrays, and document these assumptions clearly.

Let's continue to use assertions to document such assumptions in the future for ease of maintainability.