Add assertions to core classes
Assertions allow us to define assumptions about the program state so that the runtime can verify them. Assertion failures indicate possible bugs in the code.

Implementing assertions in the code allows us to affirm these assumptions to detect potential bugs.

Assertions are added to the Storage class to ensure the versatility of the file selection; the directory should be in the right directory and have the correct name, no other files in the directory, etc. We also check that the search functions in the Storage class are thorough in their search.
Similarly, assertions are added to the Duke class to ensure text fields are properly cleared after each query by the user.