Add assert
The current program do not have any assertion checks, making it
challenging to ensure the correctness of the assumption made in the application.

Assert reassures the assumption made in the program.

Adding assert to some methods to reassures assumption made. For example,
in the method list in Command, assertis used to reassure if the first
word of the message is "list".

Using Java assert statement is an efficient way to perform runtime
checks on the program which allow catching unexpected conditions and
failures early in the development process.