Storage class' updateTasksToFile method is unnecessarily changing the dates to another format. This causes clutter.
Bug is found regarding the Event tasks in that the from and to dates are being updated into the file as one string without a divider.
This causes a problem when trying to load the tasks from the file.

Let's change the storage of dates into the file without changing its DateTime format.

We can also change the from and to dates for Event tasks to be separated by a divider to allow the loadTasksFromFile to read the dates properly.