Add assertions for non-empty text
Before this commit, asserting that the text given to the parseComponent and parseCommand functions was done by checking for non-emptiness and throwing an IllegalArgumentException. While this was desirable behavior, I believe assertions are better suited for the task.

This is because this non-empty invariant should be upholded by the programmer who is building Ekud and not the user of the program. Throwing an exception might result in it getting caught by a regular try-catch and eventually being displayed to the user, which is not the purpose of this check.

Therefore, we switch to using assertions here instead of a simple IllegalArgumentException.