Improve code quality
Currently, there are a number of parts of the code that can be improved.

Hence there are changes to the code to make it easier to read.

Changes include:
* Format Javadoc for parameters.
* Break long methods into a combination of shorter methods
in Storage and Parser classes.
* Declare task type in Task interface and
remove task type declaration elsewhere.

This commit also fixes a bug where the command to list a type of task
displays all tasks without type filtering.