Refactor Command classes
Parser now is shorter and it fulfills SLAP as it
only has the functionality to return a command, the
functionality to return a string with the command is now
at the command classes.

Parser was very long and does not fulfill SLAP.

This will make the code more readable and improve the code quality.
This is part of A-CodeQuality.