Add assertions to code
Assertions are important because they help catch and identify critical
errors early in the development process, ensuring that the code behaves
as expected and preventing regression in later stages of development.

Add assertions to the Commands, Task, Todo, Event and Deadline classes
to check that values passed in are not null.

This is done to prevent when there are modifications to code, no null
classes are accidentally passed in, causing unnecessary errors.