Update A-Assertions There is no assertions in the code
Assertions are used to define assumptions about the program state so that the runtime can verify them.
Assertions serve a valuable purpose by aiding programmers
in verifying the accuracy of our assumptions and
ensuring that no invariants are breached.

We should incorporate assertions into the code in areas
where it's crucial to validate these assumptions.