Add Update Feature
Lacking in capability to modify tasks after they are created.
User has to delete an existing task if they want to change the
description or date of a task.

Introduce update feature to allow users to modify either descriptions
or date-times or both.
Specifically, for `Todo` tasks, it only allows for updates to
descriptions only.