Add friendlier syntax
Typing out the entire command can be tedious and frustrating
when dealing with many tasks at hand.

Added multiple shorthands for users to be able to
easily specify the command type without having the need
to type out the entire command.

This allows users to be able to use the functions of the
chat bot much more easily and provides for a simpler
experience.