Add assertions to the project
Assertions are not used for the Anto Chatbot.

Assertions can help us ensure that the important assumptions at various
points of the code are held true.

Let's add assert statements into various parts of the code to check that
our assumptions are true.

We need to ensure our assumptions about the code is true at various
points and we can use assert statements to do so.