Improve Code Quality
Improving code quality is important because it enhances software
reliability, maintainability, and scalability while reducing the
likelihood of bugs and costly future fixes.

Rename class and variable names to obviously display what the code is
doing, Delete unused codes is some classes, Simplify complicated
expressions, Modify long expressions, Add code comments to codes that
look slightly more complicated.

This is done so that the codes are easier to read and understand.