Assumptions at certain stages of the program are not verified when calling certain methods.
If the assumptions do not hold, the program might throw runtime errors
when a user is interacting with it.

Add assertions to determine if these assumptions hold at point of
execution of the methods.

These will highlight any cases in which the assumptions do not hold so
that attention can be drawn to bugs in the code.