Add feature to detect duplicates.
Users are able to add multiple tasks which have the same attributes.

This is redundant as having 2 identical tasks only makes the list harder
to read.

Upon adding a new task, the bot will not add the task if it already
exists and inform the user of an existing task.

This prevents duplicates from being added to the list.