Add a postpone feature
Currently there is no way to change the deadline of a Deadline task or the duration of a Event task.

Users must delete the current task and add a new task to change its details.

Let's add a feature to allow users to postpone deadlines/durations!

Users can change the details of an event more easily.