Update GUI welcome message
Paimonbot GUI shows welcome message when user inputs 'hello' command.

This causes a blank GUI to be shown when user launches the chatbot which
can be very dull.

The initialize method in MainWindow now calls the dialogContainer to create
a new dialogBox with the paimonImage and the welcome message to be printed
out when the program is ran.

Showcasing a welcome message when chatbot is launched is more vibrant compared
to just having a blank space.