Add update and clone functionality
The ability to update and clone tasks without deleting them or adding
the details all over again helps the user to improve the ease of use of
the application.

By adding the update functionality, tasks can now be updated using
simple commands such as `update 4 msg Hello World` or `update 3 Deadline
2023-10-10T12:34:56`. This updates the message and deadline for the
tasks respectively. This is done by making the Tasks more mutable and
allowing for updates to be directly done to the task.

By adding the clone functionality, users can add Tasks based on another
task without having to specify the task all over again. This may be
helpful for recurring tasks or related tasks.

These implementations are for the extension **C-Update**.