Add SortCommand class
There is no way to sort tasks by date or time.

Added a SortCommand class with methods that allow for the sorting of
deadlines and events in ascending order of date or time.

Users will most likely want to visualise which deadlines or events are
coming soon.

Let's allow the sorting of tasks with a command.