Handle Duplicate Tasks Entered
Possible to enter two same tasks with same description, date and time.

This does not make sense as the user should not have two of the same
tasks. They may have either forgotten they have entered the same task
before or mistyped.

Let's handle exceptions so that every time the user tries to add a task,
the Parser class checks if that task is already in the taskList through
an overriden equals method for each of the subclasses of Task. If a task
is in the task list, do not add it and inform the user.

Overriding the equals method is preferable as it is more maintainable
and extensible compared to hard-coding the equality check into the
method that checks against every task in the task list.