BCD-Extension C-Detectduplicates
Adding this extension is important because it improves the MYBOT system
and warns users if event has already been added.

Modify all addXXTask methods to check if there is a task of the same
name. Add methods and parameters to task list such that it
can temporarily store the task. Modify the commands so that if user
confirms that they want to add the item again, they need not retype the
task.