Improve code quality
Current code structure is not optimal for readability and there are some violations to coding standards.

In the Parser class, part of the logic in parseCommand method is extracted to various private methods for dealing with different types of command that user inputs, reduce the length of the method. Removing unnecessary import statements to reduce clutter in code and obey coding standards.

This allows reader to easily process the code.