Check for duplicates
TaskList add command: check for duplicate tasks

The add method does not check for duplicates, allowing users to add the
same task multiple times.

Adding the ability to recognise and detect duplicates allows for a
cleaner task list.

Let's add the ability to detect duplicates by checking if there is
already a task with the same description in the task list. If there is
such a task present, tell the user that there is already an existing
task with the same description in the task list.

This prevents users from adding tasks with the same description over and
over again in the task list.