Add detect duplicate feature
HelpBuddy allows users to add tasks with the same details repeatedly into the task list.

Before adding any task that user inputs, there will be a check to ensure that this task is not present in the task list.

This removes clutter arising from duplicate tasks when users want to see a list of their saved tasks.