Created a function to detect anomalies in events.
The user is able to add events, at times these events might clash.

By creating a detect anomaly function, the chatbot detects possible
event clashes and warns the user that there is a clash, and thus
does not allow the creation of another event over the same time period.

This is done through the creation of an event list class, which stores only
the events, and then compares each event's timing with the new event's
time. Doing so will give the user more control over its events.