Refactor code through extraction of checking input methods
Parser and Storage classes have some repeated lines of code (eg.
checking if String input is blank and throwing an exception).

The common behaviours throughout the classes cause code duplication.

Extracting the common behaviour into a separate method in the respective
class helps to consolidate duplicate sections.

Let's add a new method in the Parser and Storage classes that moves
the duplicated code outside of the expression.