Add assert statements to document assumptions
There are no assert statements which causes some incorrect inputs
to not trigger an exception. There is also an unused UI class
that was previously used to output responses.

Assert statements are added to document assumptions and prevent
more incorrect inputs which cause exceptions. The unused UI class
was also deleted.