Improve code quality
There are areas within the code that can be of higher quality.

Improving code quality is important in making the code more readable
and easier to understand for anyone who reads the code.

Let's improve the quality of the code in order to be able to achieve
that goal,
* by minimizing scope of variables as much as possible
* by making happy paths more prominent whenever possible

Doing this is critical to improving the quality of the code.