Debug validate methods for todo, deadline and event commands
Previous implementation did not allow for users to include the words "todo", "deadline" and "event" in their description as method used to replace those words for their respective commands.

Adjusting the validation to allow for those words to be included in the description results in better code quality and flexibility.