Update code to include assertions
Code has exceptions to handle any unsual conditions created by the
user. However, there is a lack of handlers for errors by the
programmer.

This makes the code prone to bugs due to mistakes by the programmer.
Without a handler for these potential bugs, the bot might break while
running.

Let's include assertions to account for mistakes by the programmer.

This makes the code safe by terminating the execution when there is
a possible bug detected.