Improve code quality
Code, while functional, lacks readability and is error prone. This
increases the likelihood of having bugs that are harder to detect.

Let's improve the code quality by refactoring the code.

This makes the code more readable and less prone to errors. Even if
errors are detected, it will be easier to locate the bugs as well.