Add Savable interface
The storage class takes in an ArrayList of Strings to save to file.

Does not fully abstract out the functionality to saving objects to
file as it requires the object to be converted into a string before
being passed to saveData in the Storage class.

Adding a Savable interface to be implemented by all classes that can be
saved to file allows for full abstraction.