Add assertions to classes
There are several classes that require certain invariants to be true.
For example, the TaskList class should not have an uninitialised list.
We can use assertions to check that these invariants are true.

Assertions are used to check that certain conditions are true.

Let's add additional assertions as required to make sure all invariants are
satisfied when a method executes.