Add Undo function into the chatbot to undo previous command
There is instance where the user require to revert back the command, without thinking of using the counter command

The "undo" command can be use to revert back the valid command executed previously, by automatically execute the counter
command. The stack data structure is used to store all the valid commands and the undo can be repeatedly use to revert all the way
to the fresh state.

By doing so, it improve the convenience for the user to modify their tasks storing.