Add the ability to recognize and deal with duplicate items
The user can to add duplicate task in the task list making which compromising the effectiveness of Chatty, the task management system.

Adding the ability to recognise and deal with duplicate items will prevent the list from having 2 or more of the same task.

Lets change the parent class Task to abstract and add an abstract method to check if the new Task Object is a duplicate of the curretn Task. Event, ToDo and Deadline class override this abstract method and check the different fields accordingly.

Add the ability to recognize and deal with duplicate items

The user can to add duplicate task in the task list making which compromising the effectiveness of Chatty, the task management system.

Adding the ability to recognise and deal with duplicate items will prevent the list from having 2 or more of the same task.

Lets change the parent class Task to abstract and add an abstract method to check if the new Task Object is a duplicate of the curretn Task. Event, ToDo and Deadline class override this abstract method and check the different fields accordingly.

This follows the OOP properties as it assigns the responsibility of duplicate checking to each subclass, ensuring that they manage duplicates within their respective domains.